refactor: JPA Mapping 및 api 수정작업 완료

+++ backend/src/main/java/com/dongguk/cse/naemansan/service/ImageService.java
@@ -1,8 +1,12 @@
 package com.dongguk.cse.naemansan.service;
 
 import com.dongguk.cse.naemansan.domain.Image;
+import com.dongguk.cse.naemansan.domain.User;
 import com.dongguk.cse.naemansan.domain.type.ImageUseType;
+import com.dongguk.cse.naemansan.repository.AdvertisementRepository;
 import com.dongguk.cse.naemansan.repository.ImageRepository;
+import com.dongguk.cse.naemansan.repository.ShopRepository;
+import com.dongguk.cse.naemansan.repository.UserRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -20,6 +24,9 @@ import java.util.UUID;
 @RequiredArgsConstructor
 @Transactional
 public class ImageService {
+    private final UserRepository userRepository;
+    private final ShopRepository shopRepository;
+    private final AdvertisementRepository advertisementRepository;
     private final ImageRepository imageRepository;
 
     private final String FOLDER_PATH="C:/Users/HyungJoon/Documents/0_OSSP/resources/images/";
@@ -36,38 +43,50 @@ public class ImageService {
             return "file uploaded Fail : " + filePath;
         }
 
-        Optional<Image> findImage = imageRepository.findByUserId(useId);
+        Optional<? extends Object> useObject = null;
+        switch (imageUseType) {
+            case USER -> { useObject = userRepository.findById(useId); }
+            case SHOP -> { useObject = shopRepository.findById(useId); }
+            case ADVERTISEMENT -> { useObject = advertisementRepository.findById(useId); }
+        }
+
+        Optional<Image> findImage = imageRepository.findByImageUser((User) useObject.get());
 
         if (findImage.isEmpty()) {
             imageRepository.save(Image.builder()
-                    .useId(useId)
+                    .userObject(useObject)
                     .imageUseType(imageUseType)
                     .originName(file.getOriginalFilename())
                     .uuidName(uuidImageName)
                     .type(file.getContentType())
                     .path(filePath).build());
         } else {
-            File currentFile = new File(findImage.get().getPath());
-            boolean result = currentFile.delete();
+            if (!findImage.get().getOriginName().equals("default_image.png")) {
+                File currentFile = new File(findImage.get().getPath());
+                boolean result = currentFile.delete();
+            }
 
-            findImage.get().setOriginName(file.getOriginalFilename());
-            findImage.get().setUuidName(uuidImageName);
-            findImage.get().setPath(filePath);
-            findImage.get().setType(file.getContentType());
+            findImage.get().updateImage(file.getOriginalFilename(), uuidImageName, filePath, file.getContentType());
         }
 
         return "file uploaded successfully : " + filePath;
     }
 
     public byte[] downloadImage(String fileName) throws IOException {
-        Optional<Image> image = imageRepository.findByUuidName(fileName);
+        String filePath = null;
 
-        if (image.isEmpty()) {
-            log.error("존재하지 않는 파일입니다 - UUID: {}", fileName);
-            return null;
+        if (fileName.equals("0_default_image.png")) {
+            filePath = "C:/Users/HyungJoon/Documents/0_OSSP/resources/images/0_default_image.png";
+        } else {
+            Optional<Image> image = imageRepository.findByUuidName(fileName);
+
+            if (image.isEmpty()) {
+                log.error("존재하지 않는 파일입니다 - UUID: {}", fileName);
+                return null;
+            }
+            fileName = image.get().getPath();
         }
 
-        String filePath = image.get().getPath();
         byte[] images = Files.readAllBytes(new File(filePath).toPath());
         return images;
     }
